Crate google_photoslibrary1

This documentation was generated from Photos Library crate version 1.0.9+20190702, where 20190702 is the exact revision of the photoslibrary:v1 schema built by the mako code generator v1.0.9.

Everything else about the Photos Library v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

albums

add enrichment, batch add media items, batch remove media items, create, get, list, share and unshare

media items

batch create, batch get, get, list and search

shared albums

get, join, leave and list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.albums().batch_remove_media_items(...).doit()
let r = hub.shared_albums().get(...).doit()
let r = hub.albums().batch_add_media_items(...).doit()
let r = hub.albums().add_enrichment(...).doit()
let r = hub.albums().list(...).doit()
let r = hub.albums().unshare(...).doit()
let r = hub.albums().share(...).doit()
let r = hub.albums().create(...).doit()
let r = hub.albums().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-photoslibrary1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_photoslibrary1 as photoslibrary1;
use photoslibrary1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use photoslibrary1::PhotosLibrary;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = PhotosLibrary::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.albums().list()
             .page_token("sed")
             .page_size(-85)
             .exclude_non_app_created_data(true)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AddEnrichmentToAlbumRequest
Request to add an enrichment to a specific album at a specific position.

AddEnrichmentToAlbumResponse
The enrichment item that's created.

Album
Representation of an album in Google Photos. Albums are containers for media items. If an album has been shared by the application, it contains an extra shareInfo property.

AlbumAddEnrichmentCall
Adds an enrichment at a specified position in a defined album.

AlbumBatchAddMediaItemCall
Adds one or more media items in a user's Google Photos library to an album. The media items and albums must have been created by the developer via the API.

AlbumBatchRemoveMediaItemCall
Removes one or more media items from a specified album. The media items and the album must have been created by the developer via the API.

AlbumCreateCall
Creates an album in a user's Google Photos library.

AlbumGetCall
Returns the album based on the specified albumId. The albumId must be the ID of an album owned by the user or a shared album that the user has joined.

AlbumListCall
Lists all albums shown to a user in the Albums tab of the Google Photos app.

AlbumMethods
A builder providing access to all methods supported on album resources. It is not used directly, but through the PhotosLibrary hub.

AlbumPosition
Specifies a position in an album.

AlbumShareCall
Marks an album as shared and accessible to other users. This action can only be performed on albums which were created by the developer via the API.

AlbumUnshareCall
Marks a previously shared album as private. This means that the album is no longer shared and all the non-owners will lose access to the album. All non-owner content will be removed from the album. If a non-owner has previously added the album to their library, they will retain all photos in their library. This action can only be performed on albums which were created by the developer via the API.

BatchAddMediaItemsToAlbumRequest
Request to add media items to an album.

BatchAddMediaItemsToAlbumResponse
Response for adding media items to an album.

BatchCreateMediaItemsRequest
Request to create one or more media items in a user's Google Photos library. If an albumid is specified, the media items are also added to that album. albumPosition is optional and can only be specified if an albumId is set.

BatchCreateMediaItemsResponse
List of media items created.

BatchGetMediaItemsResponse
Response to retrieve a list of media items.

BatchRemoveMediaItemsFromAlbumRequest
Request to remove a list of media items from an album.

BatchRemoveMediaItemsFromAlbumResponse
Response for successfully removing all specified media items from the album.

Chunk

ContentFilter
This filter allows you to return media items based on the content type.

ContentRange
Implements the Content-Range header, for serialization only

ContributorInfo
Information about the user who added the media item. Note that this information is included only if the media item is within a shared album created by your app and you have the sharing scope.

CreateAlbumRequest
Request to create an album in Google Photos.

Date
Represents a whole calendar date. The day may be 0 to represent a year and month where the day isn't significant, such as a whole calendar month. The month may be 0 to represent a a day and a year where the month isn't signficant, like when you want to specify the same day in every month of a year or a specific year. The year may be 0 to represent a month and day independent of year, like an anniversary date.

DateFilter
This filter defines the allowed dates or date ranges for the media returned. It's possible to pick a set of specific dates and a set of date ranges.

DateRange
Defines a range of dates. Both dates must be of the same format. For more information, see Date

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

EnrichmentItem
An enrichment item.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

FeatureFilter
This filter defines the features that the media items should have.

Filters
Filters that can be applied to a media item search. If multiple filter options are specified, they're treated as AND with each other.

JoinSharedAlbumRequest
Request to join a shared album on behalf of the user. This uses a shareToken which can be acquired via the shareAlbum or listSharedAlbums calls.

JoinSharedAlbumResponse
Response to successfully joining the shared album on behalf of the user.

JsonServerError
A utility type which can decode a server response that indicates error

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

LeaveSharedAlbumRequest
Request to leave a shared album on behalf of the user. This uses a shareToken which can be acquired via the or listSharedAlbums or getAlbum calls.

LeaveSharedAlbumResponse
Response to successfully leaving the shared album on behalf of the user.

ListAlbumsResponse
List of albums requested.

ListMediaItemsResponse
List of all media items from the user's Google Photos library.

ListSharedAlbumsResponse
List of shared albums requested.

Location
Represents a physical location.

LocationEnrichment
An enrichment containing a single location.

MapEnrichment
An enrichment containing a map, showing origin and destination locations.

MediaItem
Representation of a media item (such as a photo or video) in Google Photos.

MediaItemBatchCreateCall
Creates one or more media items in a user's Google Photos library.

MediaItemBatchGetCall
Returns the list of media items for the specified media item identifiers. Items are returned in the same order as the supplied identifiers.

MediaItemGetCall
Returns the media item for the specified media item identifier.

MediaItemListCall
List all media items from a user's Google Photos library.

MediaItemMethods
A builder providing access to all methods supported on mediaItem resources. It is not used directly, but through the PhotosLibrary hub.

MediaItemResult
Result of retrieving a media item.

MediaItemSearchCall
Searches for media items in a user's Google Photos library. If no filters are set, then all media items in the user's library are returned. If an album is set, all media items in the specified album are returned. If filters are specified, media items that match the filters from the user's library are listed. If you set both the album and the filters, the request results in an error.

MediaMetadata
Metadata for a media item.

MediaTypeFilter
This filter defines the type of media items to be returned, for example, videos or photos. All the specified media types are treated as an OR when used together.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NewEnrichmentItem
A new enrichment item to be added to an album, used by the albums.addEnrichment call.

NewMediaItem
New media item that's created in a user's Google Photos account.

NewMediaItemResult
Result of creating a new media item.

Photo
Metadata that is specific to a photo, such as, ISO, focal length and exposure time. Some of these fields may be null or not included.

PhotosLibrary
Central instance to access all PhotosLibrary related resource activities

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SearchMediaItemsRequest
Request to search for media items in a user's library.

SearchMediaItemsResponse
List of media items that match the search parameters.

ServerError

ServerMessage

ShareAlbumRequest
Request to make an album shared in Google Photos.

ShareAlbumResponse
Response to successfully sharing an album.

ShareInfo
Information about albums that are shared. This information is only included if you created the album, it is shared and you have the sharing scope.

SharedAlbumGetCall
Returns the album based on the specified shareToken.

SharedAlbumJoinCall
Joins a shared album on behalf of the Google Photos user.

SharedAlbumLeaveCall
Leaves a previously-joined shared album on behalf of the Google Photos user. The user must not own this album.

SharedAlbumListCall
Lists all shared albums available in the Sharing tab of the user's Google Photos app.

SharedAlbumMethods
A builder providing access to all methods supported on sharedAlbum resources. It is not used directly, but through the PhotosLibrary hub.

SharedAlbumOptions
Options that control the sharing of an album.

SimpleMediaItem
A simple media item to be created in Google Photos via an upload token.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

TextEnrichment
An enrichment containing text.

UnshareAlbumRequest
Request to unshare a shared album in Google Photos.

UnshareAlbumResponse
Response of a successful unshare of a shared album.

Video
Metadata that is specific to a video, for example, fps and processing status. Some of these fields may be null or not included.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.