Module google_photoslibrary1::api

Structs

AddEnrichmentToAlbumRequest
Request to add an enrichment to a specific album at a specific position.

AddEnrichmentToAlbumResponse
The enrichment item that's created.

Album
Representation of an album in Google Photos. Albums are containers for media items. If an album has been shared by the application, it contains an extra shareInfo property.

AlbumAddEnrichmentCall
Adds an enrichment at a specified position in a defined album.

AlbumBatchAddMediaItemCall
Adds one or more media items in a user's Google Photos library to an album. The media items and albums must have been created by the developer via the API. Media items are added to the end of the album. If multiple media items are given, they are added in the order specified in this call. Each album can contain up to 20,000 media items. Only media items that are in the user's library can be added to an album. For albums that are shared, the album must either be owned by the user or the user must have joined the album as a collaborator. Partial success is not supported. The entire request will fail if an invalid media item or album is specified.

AlbumBatchRemoveMediaItemCall
Removes one or more media items from a specified album. The media items and the album must have been created by the developer via the API. For albums that are shared, this action is only supported for media items that were added to the album by this user, or for all media items if the album was created by this user. Partial success is not supported. The entire request will fail and no action will be performed on the album if an invalid media item or album is specified.

AlbumCreateCall
Creates an album in a user's Google Photos library.

AlbumGetCall
Returns the album based on the specified albumId. The albumId must be the ID of an album owned by the user or a shared album that the user has joined.

AlbumListCall
Lists all albums shown to a user in the Albums tab of the Google Photos app.

AlbumMethods
A builder providing access to all methods supported on album resources. It is not used directly, but through the PhotosLibrary hub.

AlbumPatchCall
Update the album with the specified id. Only the id, title and cover_photo_media_item_id fields of the album are read. The album must have been created by the developer via the API and must be owned by the user.

AlbumPosition
Specifies a position in an album.

AlbumShareCall
Marks an album as shared and accessible to other users. This action can only be performed on albums which were created by the developer via the API.

AlbumUnshareCall
Marks a previously shared album as private. This means that the album is no longer shared and all the non-owners will lose access to the album. All non-owner content will be removed from the album. If a non-owner has previously added the album to their library, they will retain all photos in their library. This action can only be performed on albums which were created by the developer via the API.

BatchAddMediaItemsToAlbumRequest
Request to add media items to an album.

BatchAddMediaItemsToAlbumResponse
Response for adding media items to an album.

BatchCreateMediaItemsRequest
Request to create one or more media items in a user's Google Photos library. If an albumid is specified, the media items are also added to that album. albumPosition is optional and can only be specified if an albumId is set.

BatchCreateMediaItemsResponse
List of media items created.

BatchGetMediaItemsResponse
Response to retrieve a list of media items.

BatchRemoveMediaItemsFromAlbumRequest
Request to remove a list of media items from an album.

BatchRemoveMediaItemsFromAlbumResponse
Response for successfully removing all specified media items from the album.

ContentFilter
This filter allows you to return media items based on the content type. It's possible to specify a list of categories to include, and/or a list of categories to exclude. Within each list, the categories are combined with an OR. The content filter includedContentCategories: [c1, c2, c3] would get media items that contain (c1 OR c2 OR c3). The content filter excludedContentCategories: [c1, c2, c3] would NOT get media items that contain (c1 OR c2 OR c3). You can also include some categories while excluding others, as in this example: includedContentCategories: [c1, c2], excludedContentCategories: [c3, c4] The previous example would get media items that contain (c1 OR c2) AND NOT (c3 OR c4). A category that appears in includedContentategories must not appear in excludedContentCategories.

ContributorInfo
Information about the user who added the media item. Note that this information is included only if the media item is within a shared album created by your app and you have the sharing scope.

CreateAlbumRequest
Request to create an album in Google Photos.

Date
Represents a whole calendar date. Set day to 0 when only the month and year are significant, for example, all of December 2018. Set day and month to 0 if only the year is significant, for example, the entire of 2018. Set year to 0 when only the day and month are significant, for example, an anniversary or birthday. Unsupported: Setting all values to 0, only month to 0, or both day and year to 0 at the same time.

DateFilter
This filter defines the allowed dates or date ranges for the media returned. It's possible to pick a set of specific dates and a set of date ranges.

DateRange
Defines a range of dates. Both dates must be of the same format. For more information, see Date.

EnrichmentItem
An enrichment item.

FeatureFilter
This filter defines the features that the media items should have.

Filters
Filters that can be applied to a media item search. If multiple filter options are specified, they're treated as AND with each other.

JoinSharedAlbumRequest
Request to join a shared album on behalf of the user. This uses a shareToken which can be acquired via the shareAlbum or listSharedAlbums calls.

JoinSharedAlbumResponse
Response to successfully joining the shared album on behalf of the user.

LatLng
An object that represents a latitude/longitude pair. This is expressed as a pair of doubles to represent degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

LeaveSharedAlbumRequest
Request to leave a shared album on behalf of the user. This uses a shareToken which can be acquired via the or listSharedAlbums or getAlbum calls.

LeaveSharedAlbumResponse
Response to successfully leaving the shared album on behalf of the user.

ListAlbumsResponse
List of albums requested.

ListMediaItemsResponse
List of all media items from the user's Google Photos library.

ListSharedAlbumsResponse
List of shared albums requested.

Location
Represents a physical location.

LocationEnrichment
An enrichment containing a single location.

MapEnrichment
An enrichment containing a map, showing origin and destination locations.

MediaItem
Representation of a media item (such as a photo or video) in Google Photos.

MediaItemBatchCreateCall
Creates one or more media items in a user's Google Photos library. This is the second step for creating a media item. For details regarding Step 1, uploading the raw bytes to a Google Server, see Uploading media. This call adds the media item to the library. If an album id is specified, the call adds the media item to the album too. Each album can contain up to 20,000 media items. By default, the media item will be added to the end of the library or album. If an album id and position are both defined, the media item is added to the album at the specified position. If the call contains multiple media items, they're added at the specified position. If you are creating a media item in a shared album where you are not the owner, you are not allowed to position the media item. Doing so will result in a BAD REQUEST error.

MediaItemBatchGetCall
Returns the list of media items for the specified media item identifiers. Items are returned in the same order as the supplied identifiers.

MediaItemGetCall
Returns the media item for the specified media item identifier.

MediaItemListCall
List all media items from a user's Google Photos library.

MediaItemMethods
A builder providing access to all methods supported on mediaItem resources. It is not used directly, but through the PhotosLibrary hub.

MediaItemPatchCall
Update the media item with the specified id. Only the id and description fields of the media item are read. The media item must have been created by the developer via the API and must be owned by the user.

MediaItemResult
Result of retrieving a media item.

MediaItemSearchCall
Searches for media items in a user's Google Photos library. If no filters are set, then all media items in the user's library are returned. If an album is set, all media items in the specified album are returned. If filters are specified, media items that match the filters from the user's library are listed. If you set both the album and the filters, the request results in an error.

MediaMetadata
Metadata for a media item.

MediaTypeFilter
This filter defines the type of media items to be returned, for example, videos or photos. All the specified media types are treated as an OR when used together.

NewEnrichmentItem
A new enrichment item to be added to an album, used by the albums.addEnrichment call.

NewMediaItem
New media item that's created in a user's Google Photos account.

NewMediaItemResult
Result of creating a new media item.

Photo
Metadata that is specific to a photo, such as, ISO, focal length and exposure time. Some of these fields may be null or not included.

PhotosLibrary
Central instance to access all PhotosLibrary related resource activities

SearchMediaItemsRequest
Request to search for media items in a user's library. If the album id is specified, this call will return the list of media items in the album. If neither filters nor album id are specified, this call will return all media items in a user's Google Photos library. If filters are specified, this call will return all media items in the user's library that fulfill the filter criteria. Filters and album id must not both be set, as this will result in an invalid request.

SearchMediaItemsResponse
List of media items that match the search parameters.

ShareAlbumRequest
Request to make an album shared in Google Photos.

ShareAlbumResponse
Response to successfully sharing an album.

ShareInfo
Information about albums that are shared. This information is only included if you created the album, it is shared and you have the sharing scope.

SharedAlbumGetCall
Returns the album based on the specified shareToken.

SharedAlbumJoinCall
Joins a shared album on behalf of the Google Photos user.

SharedAlbumLeaveCall
Leaves a previously-joined shared album on behalf of the Google Photos user. The user must not own this album.

SharedAlbumListCall
Lists all shared albums available in the Sharing tab of the user's Google Photos app.

SharedAlbumMethods
A builder providing access to all methods supported on sharedAlbum resources. It is not used directly, but through the PhotosLibrary hub.

SharedAlbumOptions
Options that control the sharing of an album.

SimpleMediaItem
A simple media item to be created in Google Photos via an upload token.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

TextEnrichment
An enrichment containing text.

UnshareAlbumRequest
Request to unshare a shared album in Google Photos.

UnshareAlbumResponse
Response of a successful unshare of a shared album.

Video
Metadata that is specific to a video, for example, fps and processing status. Some of these fields may be null or not included.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.