Struct google_photoslibrary1::AlbumMethods

pub struct AlbumMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on album resources. It is not used directly, but through the PhotosLibrary hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_photoslibrary1 as photoslibrary1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use photoslibrary1::PhotosLibrary;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = PhotosLibrary::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_enrichment(...)`, `batch_add_media_items(...)`, `batch_remove_media_items(...)`, `create(...)`, `get(...)`, `list(...)`, `share(...)` and `unshare(...)`
// to build up your call.
let rb = hub.albums();

Methods

impl<'a, C, A> AlbumMethods<'a, C, A>

pub fn batch_remove_media_items(
&self,
request: BatchRemoveMediaItemsFromAlbumRequest,
album_id: &str
) -> AlbumBatchRemoveMediaItemCall<'a, C, A>

Create a builder to help you perform the following task:

Removes one or more media items from a specified album. The media items and the album must have been created by the developer via the API.

For albums that are shared, this action is only supported for media items that were added to the album by this user, or for all media items if the album was created by this user.

Partial success is not supported. The entire request will fail and no action will be performed on the album if an invalid media item or album is specified.

Arguments

request - No description provided.

albumId - Identifier of the Album that the media items are to be removed from.

pub fn batch_add_media_items(
&self,
request: BatchAddMediaItemsToAlbumRequest,
album_id: &str
) -> AlbumBatchAddMediaItemCall<'a, C, A>

Create a builder to help you perform the following task:

Adds one or more media items in a user's Google Photos library to an album. The media items and albums must have been created by the developer via the API.

Media items are added to the end of the album. If multiple media items are given, they are added in the order specified in this call.

Each album can contain up to 20,000 media items.

Only media items that are in the user's library can be added to an album. For albums that are shared, the album must either be owned by the user or the user must have joined the album as a collaborator.

Partial success is not supported. The entire request will fail if an invalid media item or album is specified.

Arguments

request - No description provided.

albumId - Identifier of the Album that the media items are added to.

pub fn add_enrichment(
&self,
request: AddEnrichmentToAlbumRequest,
album_id: &str
) -> AlbumAddEnrichmentCall<'a, C, A>

Create a builder to help you perform the following task:

Adds an enrichment at a specified position in a defined album.

Arguments

request - No description provided.

albumId - Identifier of the album where the enrichment is to be added.

pub fn list(&self) -> AlbumListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all albums shown to a user in the Albums tab of the Google Photos app.

pub fn unshare(
&self,
request: UnshareAlbumRequest,
album_id: &str
) -> AlbumUnshareCall<'a, C, A>

Create a builder to help you perform the following task:

Marks a previously shared album as private. This means that the album is no longer shared and all the non-owners will lose access to the album. All non-owner content will be removed from the album. If a non-owner has previously added the album to their library, they will retain all photos in their library. This action can only be performed on albums which were created by the developer via the API.

Arguments

request - No description provided.

albumId - Identifier of the album to be unshared. This album id must belong to an album created by the developer.

pub fn share(
&self,
request: ShareAlbumRequest,
album_id: &str
) -> AlbumShareCall<'a, C, A>

Create a builder to help you perform the following task:

Marks an album as shared and accessible to other users. This action can only be performed on albums which were created by the developer via the API.

Arguments

request - No description provided.

albumId - Identifier of the album to be shared. This albumId must belong to an album created by the developer.

pub fn create(&self, request: CreateAlbumRequest) -> AlbumCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an album in a user's Google Photos library.

Arguments

request - No description provided.

pub fn get(&self, album_id: &str) -> AlbumGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the album based on the specified albumId. The albumId must be the ID of an album owned by the user or a shared album that the user has joined.

Arguments

albumId - Identifier of the album to be requested.